refactor: relocate drift numerical helpers
