+++ src/quantem/imaging/drift/__init__.py
@@ -1,9 +1,11 @@
 from quantem.imaging.drift.correction import DriftCorrection as DriftCorrection
-from quantem.imaging.drift.core.warping import (
-    DriftInterpolator as DriftInterpolator,
+from quantem.imaging.drift.apply import (
     _bounded_sine_sigmoid_torch as _bounded_sine_sigmoid_torch,
     _fourier_crop_torch as _fourier_crop_torch,
     bounded_sine_sigmoid as bounded_sine_sigmoid,
+)
+from quantem.imaging.drift.core.knots import (
+    DriftInterpolator as DriftInterpolator,
 )
 
 DriftCorrection.__module__ = __name__
