+++ src/quantem/imaging/drift/core/strip.py
@@ -1,0 +1,1 @@
+"""Package boundary for the strip-correction stage added in a later PR."""
